requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-debug-toolbar
python-dotenv
gunicorn